Add files via upload
Added Shop
-When clicking on shop, creates shop sprites (upgrade icons)
-If shop is already open, destroys the shop sprites
Added a new package: Upgrades
-Contains all the upgrades to be implemented
-Probably make an interface file so I don't have to copy paste doClick x amount of times

+++ src/main/java/com/johnnyong/android/gamedevbumperhero/HeroSprite.kt
@@ -17,7 +17,7 @@ class HeroSprite(
     private var xVelocity = newHeroVelocity
 
     override fun draw(canvas: Canvas){
-        canvas.drawBitmap(playerImage, heroX, heroY.toFloat(), null)
+        canvas.drawBitmap(playerImage, heroX, heroY, null)
     }
 
     override fun update() {
